:bug: Fix 'currentPurchase' reappering on iOS
When the user visits the app again after subscribing, 'currentPurchase' comes back as not null. That caused the app to attempt to initiate a new subscription, which did not work anyway. This commit fixes that by creating a 'hasRequestedSubscription' state, which means the purchase is only going to be sent if the user has clicked the 'subscribe' button.