:bug: Fix route recalculation issue
Since each coordinate (latitude and longitude) had its own state, as soon as one of them was set, the route would be calculated with the old value of the second one, which had not been set yet. This commit replaces the previous states with ones encompassing both latitude and longitude values.